Modularize structure.sequence into a submodule to simplify code
This modularized the logic sufficiently that the logic in
PeptideSequence could be split between a core base class and
mixins that manipulate the properties of that base class.

This let me re-write the core base class in C for substantially
improved performance during instantiation, parsing, and interop
with other extension modules, particularly fragmentation_strategy.

During the optimization of fragmentation_strategy, I introduced
the count_table C extension which implements a dict-like C struct
that uses arbitrary Python objects (refcounted but stored as PyObject*)
and C longs as values. This gives nearly the same performance as a
Python Dict (a little slower because it can't optimize away calls to
hash), while providing better count arithmetic performance and
maintaining count types. This type could be useful elsewhere.